# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Dustid Contact Verification API
  version: 1.0.0
  description: Simple mock API for OTP verification and contact lookup.

servers:
  - url: http://localhost:3000
    description: Local development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Message:
      type: object
      properties:
        message:
          type: string
        otp:
          type: string
          description: Development-only OTP preview
    Address:
      type: object
      properties:
        address1:
          type: string
        address2:
          type: string
        city:
          type: string
        province:
          type: string
        zip:
          type: string
        country:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
    Friend:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        address:
          $ref: '#/components/schemas/Address'

paths:
  /:
    get:
      summary: Home route
      description: Basic check if API is running
      responses:
        '200':
          description: Welcome page
          content:
            text/html:
              schema:
                type: string

  /verify:
    post:
      summary: Generate OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
              properties:
                phoneNumber:
                  type: string
                  example: "254712345678"
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Missing phone number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Phone number not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /validate-otp:
    post:
      summary: Validate OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
                - otp
              properties:
                phoneNumber:
                  type: string
                  example: "254712345678"
                otp:
                  type: string
                  example: "123456"
      responses:
        '200':
          description: OTP validated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: OTP validated successfully.
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          description: Missing phone number or OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /friends/{friendId}:
    get:
      summary: Get a friend's full details including address
      security:
        - bearerAuth: []
      parameters:
        - name: friendId
          in: path
          required: true
          schema:
            type: string
          example: friend101
      responses:
        '200':
          description: Friend object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Friend'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          description: Friend not in your list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /search:
    post:
      summary: Search friends
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
              properties:
                phoneNumber:
                  type: string
                  example: "254712345678"
                query:
                  type: string
                  description: Optional search term
                  example: "bob"
      responses:
        '200':
          description: List of friends
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        email:
                          type: string
        '400':
          description: Missing phone number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /profile:
    post:
      summary: Get contact profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
              properties:
                phoneNumber:
                  type: string
                  example: "254712345678"
      responses:
        '200':
          description: Contact object
          content:
            application/json:
              schema:
                type: object
                properties:
                  contact:
                    type: object
                    properties:
                      id:
                        type: string
                      phoneNumber:
                        type: string
                      name:
                        type: string
                      friends:
                        type: array
                        items:
                          $ref: '#/components/schemas/Friend'
        '400':
          description: Missing phone number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'